V1/requirements.txt
tensorflow>=2.8.0
opencv-python
matplotlib
numpy
pandas
tqdm
scikit-learn
PyYAML
kaggle
